Enhancement for using yii2fullcalendar with pjax
When using the fullcalendar widget with pjax, if the user clicks the browser back button to go back to a page/view with the fullcalendar widget, the fullcalendar javascript is fired again, and the calendar becomes rendered twice.  This enhancement to the yii2fullcalendar widget addresses this issue by clearing/emptying the calendar container before rendering a new instance of the calendar.  There is probably a better/cleaner way to address this issue, however this approach works and should work without negatively affecting other possible widget use cases.